+++ specs/jpg.spec.ts
@@ -2,12 +2,7 @@ import imageSize from '../src/index';
 import toAscii from './toAscii';
 import detector from '../src/detectType';
 import setupView from './setupView'
-import toHexadecimal from '../src/toHexadecimal';
 import validateJPG from '../src/types/validateJPG';
-import readJpgBlockLength from '../src/types/readJpgBlockLength';
-import locateZeroTerminator from '../src/types/locateZeroTerminator'
-import { readUInt16BE, readUInt32BE } from '../src/readUInt';
-import findNonZeroCharacter from '../src/types/findNonZeroCharacter';
 
 describe('jpg test files', () => {
   // Inputs - fileName, path, 
@@ -105,66 +100,6 @@ describe('jpg test files', () => {
     test('imageType', () => {
       const type = detector(view, toAscii);
       expect(type).toBe(imageType)
-    })
-
-    test.skip('getBlockLength', () => {
-      // skip 4 bytes 
-      let offset = 0
-      const firstMarker = readUInt16BE(view, offset);
-      expect(firstMarker).toBe(0xffd8); // SOI - always
-      offset += 2;
-
-      const secondMarker = readUInt16BE(view, offset);
-      expect(secondMarker).toBe(markerValue); // APP1
-
-      // https://bitmiracle.github.io/libjpeg.net/help/articles/KB/special-markers.html
-      // at most Keep in mind that at most 65533 bytes can be put into one marker, but you can have as many markers as you like.      
-
-      // https://en.wikipedia.org/wiki/JPEG
-      // (The length includes the two bytes for the length, but not the two bytes for the marker.)
-
-      offset += 2;
-      // length of marker from len position HERE
-
-      const blockLength = readJpgBlockLength(view, offset);
-      expect(blockLength).toEqual(4)
-
-      let markerOffset = offset + 2;
-      const identiferEnd = locateZeroTerminator(view, markerOffset, blockLength);
-      const identifier = toAscii(view, markerOffset, identiferEnd);
-      expect(identifier).toBe(markerId);
-
-      // match identifier with second marker => solution
-        // APP1 with EXIF 
-        // APP0 with JFIF
-        // APP0 with JFXX
-        // else skip
-
-      markerOffset = findNonZeroCharacter(view, identiferEnd, blockLength);
-      // PROBABLY SKIP THESE SECTIONS
-      const majorVersion = view.getUint8(markerOffset);
-      markerOffset += 1;
-      const minorVersion = view.getUint8(markerOffset);
-      markerOffset += 1;
-      const densityUnits = view.getUint8(markerOffset);
-      markerOffset += 1;
-      const xDensity = readUInt16BE(view, offset);
-      markerOffset += 2;
-      const yDensity = readUInt16BE(view, offset);
-      markerOffset += 2;
-      const xThumbnail = view.getUint8(markerOffset);
-      markerOffset += 1;
-      const yThumbnail = view.getUint8(markerOffset);
-      markerOffset += 1;
-
-      // uncompressed 24bit-color thumbnail r0, g0, b0 .. ... Rn-1, Gn-1, Bn-1
-      const thumbnailSize = 3 * xThumbnail * yThumbnail 
-      markerOffset += thumbnailSize;
-
-      // next marker please
-      offset += blockLength;
-      const thirdMarker = readUInt16BE(view, offset);
-      expect(thirdMarker).toBe(0xffc0);
     })
 
     describe('result', () => {
